optimizing APY calls a bit
