+++ src/components/FarmSuggestionCard/index.tsx
@@ -1,6 +1,6 @@
 import { Pair, TokenAmount } from 'uniswap-xdai-sdk'
 import { darken } from 'polished'
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
 import { Text } from 'rebass'
 import styled from 'styled-components'
@@ -16,12 +16,12 @@ import { RowBetween, RowFixed } from '../Row'
 import { useTokenBalance } from '../../state/wallet/hooks'
 import { useActiveWeb3React } from '../../hooks'
 import { ExternalLink } from '../../theme'
-import { useBlockNumber } from '../../state/application/hooks'
-import { fetchPrice, useAPY, useStakedTVL } from '../../hooks/Price'
+import { useAPY, useStakedTVL } from '../../hooks/Price'
 import { useStakedAmount } from '../../data/Staked'
-import { BigNumber } from '@ethersproject/bignumber'
 import { useTotalSupply } from '../../data/TotalSupply'
 import { FarmablePool } from '../../bao/lib/constants'
+import { BigNumber } from '@ethersproject/bignumber'
+
 
 export const FixedHeightRow = styled(RowBetween)`
   height: 24px;
@@ -43,14 +43,20 @@ export const BalanceText = styled(Text)`
 interface PositionCardProps {
   pair: Pair
   farmablePool: FarmablePool
+  baoPriceUsd: BigNumber
   unstakedLPAmount?: TokenAmount | undefined | null
   showUnwrapped?: boolean
   border?: string
 }
 
-export function FarmSuggestionCard({ pair, farmablePool, showUnwrapped = true, border }: PositionCardProps) {
+export function FarmSuggestionCard({
+  pair,
+  farmablePool,
+  baoPriceUsd,
+  showUnwrapped = true,
+  border
+}: PositionCardProps) {
   const { account } = useActiveWeb3React()
-  const block = useBlockNumber()
 
   const currency0 = showUnwrapped ? pair.token0 : unwrappedToken(pair.token0)
   const currency1 = showUnwrapped ? pair.token1 : unwrappedToken(pair.token1)
@@ -74,13 +80,6 @@ export function FarmSuggestionCard({ pair, farmablePool, showUnwrapped = true, b
   const totalSupply = useTotalSupply(pair.liquidityToken)
 
   const allStakedTVL = useStakedTVL(farmablePool, stakedAmount, totalSupply)
-
-  const [baoPriceUsd, setBaoPriceUsd] = useState<BigNumber>(BigNumber.from(0))
-  useEffect(() => {
-    fetchPrice()
-      .then(apy => setBaoPriceUsd(apy))
-      .catch(() => setBaoPriceUsd(BigNumber.from(0)))
-  }, [block])
 
   const apy = useAPY(farmablePool, baoPriceUsd, allStakedTVL)
